+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.8
+
+* Added new abstraction `StreamResultMeta` to support passing additional info with `StreamResult`
+* Slight refactoring of the old code
+
 ## 0.0.6
 * Updated VGA version to `^0.0.7`, removed `flutter` environment requirements from `pubspec.yaml`